requirements.txt
sentence-transformers 
faiss-cpu 
streamlit 
requests 
tiktoken 
PyMuPDF 
pdfplumber 
markdown2